requirements.txt
numpy
openai
pynput
python-dotenv
toml